+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 6.0.1 (pending)
 
 * dumped memory quota handling, kubernetes is responsible for that
+* dumped environment field - it's too container centric now that we have kubernetes; use 'sc ssh' instead
 * server startup output with image version
 * readiness probe for stool server
 * update to junit 5.7.0